Add display user information

+++ user.dart
@@ -1,11 +1,11 @@
 class User {
   final String name;
   final int age;
-  final double _score;
+  final int _score;
 
-  double get score => _score;
+  int get score => _score;
 
-  User({required this.name, required this.age, required double score})
+  User({required this.name, required this.age, required int score})
       : _score = score;
 
   @override
